requirements.txt
anthropic>=0.40.0
feedparser>=6.0.11
requests>=2.32.0
python-dotenv>=1.0.1
beautifulsoup4>=4.12.3
